fixed dropdown menu and moved axios to UrlDataSection
