+++ app/src/components/FormElement/FormElement.js
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import api from "../../api/api";
+import React from "react";
 import {
   StyledForm,
   StyledButton,
@@ -8,54 +7,19 @@ import {
   ErrorParagraph,
 } from "./FormElement.styles";
 
-export const FormElement = ({ linksList, setLinksList }) => {
-  const [url, setUrl] = useState("");
-  const [isVisible, setIsVisible] = useState(false);
-  const [isVisible2, setIsVisible2] = useState(false);
-
-  useEffect(() => {
-    if (JSON.parse(localStorage.getItem("list")) === null) {
-      setLinksList([]);
-    } else {
-      setLinksList(JSON.parse(localStorage.getItem("list")));
-    }
-  }, []);
-
+export const FormElement = ({
+  url,
+  shortenLink,
+  setUrl,
+  isVisible,
+  isVisible2,
+  setIsVisible,
+  setIsVisible2,
+}) => {
   const handleInput = (e) => {
     setUrl(e.target.value);
     setIsVisible(false);
     setIsVisible2(false);
-  };
-
-  const shortenLink = async (e) => {
-    e.preventDefault();
-    if (url.trim().length > 0) {
-      try {
-        const response = await api.get(`/shorten?url=${url}`);
-        setLinksList([
-          ...linksList,
-          { longLink: url, shortLink: response.data.result.short_link },
-        ]);
-        localStorage.setItem(
-          "list",
-          JSON.stringify([
-            ...linksList,
-            { longLink: url, shortLink: response.data.result.short_link },
-          ])
-        );
-      } catch (err) {
-        if (err.response) {
-          console.log(err.response.data);
-          console.log(err.response.status);
-        } else {
-          console.log(`Error: ${err.message}`);
-        }
-      }
-      setUrl("");
-    } else {
-      setIsVisible(true);
-      setIsVisible2(true);
-    }
   };
 
   return (
